move the `TaskToCompute` test to `test_tasks`
