+++ tests/message/test_messages.py
@@ -6,11 +6,8 @@ import unittest.mock as mock
 import uuid
 
 from golem_messages import datastructures
-from golem_messages import exceptions
 from golem_messages import message
 from golem_messages import shortcuts
-
-from tests import factories
 
 
 class InitializationTestCase(unittest.TestCase):
@@ -352,30 +349,3 @@ class MessagesTestCase(unittest.TestCase):
             ['options', options],
         ]
         self.assertEqual(expected, msg.slots())
-
-    def test_task_to_compute_basic(self):
-        ttc = factories.TaskToComputeFactory()
-        serialized = shortcuts.dump(ttc, None, None)
-        msg = shortcuts.load(serialized, None, None)
-        self.assertIsInstance(msg, message.tasks.TaskToCompute)
-
-    def test_task_to_compute_validate_compute_task_def(self):
-        requestor_id = 'such as epidemiology'
-        # Shoudn't raise
-        message.TaskToCompute(slots=(
-            ('requestor_id', requestor_id),
-        ))
-
-        compute_task_def = message.ComputeTaskDef(
-            task_owner={'key': requestor_id}
-        )
-        message.TaskToCompute(slots=(
-            ('requestor_id', requestor_id),
-            ('compute_task_def', compute_task_def),
-        ))
-
-        with self.assertRaises(exceptions.FieldError):
-            message.TaskToCompute(slots=(
-                ('requestor_id', 'staple of research'),
-                ('compute_task_def', compute_task_def),
-            ))
