+++ tests/message/test_tasks.py
@@ -2,7 +2,9 @@
 import unittest
 import unittest.mock as mock
 
+from golem_messages import exceptions
 from golem_messages import message
+from golem_messages import shortcuts
 
 from tests import factories
 
@@ -68,3 +70,31 @@ class TasksTest(unittest.TestCase):
         self.assertEqual(expected, msg.slots())
         self.assertIsInstance(msg.force_get_task_result_failed,
                               message.concents.ForceGetTaskResultFailed)
+
+class TaskToComputeTest(unittest.TestCase):
+    def test_task_to_compute_basic(self):
+        ttc = factories.TaskToComputeFactory()
+        serialized = shortcuts.dump(ttc, None, None)
+        msg = shortcuts.load(serialized, None, None)
+        self.assertIsInstance(msg, message.tasks.TaskToCompute)
+
+    def test_task_to_compute_validate_compute_task_def(self):
+        requestor_id = 'such as epidemiology'
+        # Shoudn't raise
+        message.TaskToCompute(slots=(
+            ('requestor_id', requestor_id),
+        ))
+
+        compute_task_def = message.ComputeTaskDef(
+            task_owner={'key': requestor_id}
+        )
+        message.TaskToCompute(slots=(
+            ('requestor_id', requestor_id),
+            ('compute_task_def', compute_task_def),
+        ))
+
+        with self.assertRaises(exceptions.FieldError):
+            message.TaskToCompute(slots=(
+                ('requestor_id', 'staple of research'),
+                ('compute_task_def', compute_task_def),
+            ))
